Reviewed, cleaned up and documented createAuth
